Further HostPort refactorings to decouple the parsing from the utility classes with the introduction of HostPort#parse methods
